requirements.txt
ethereum==1.3.6
gevent==1.1.0
pyethapp==1.2.3
pymongo==3.3.0
python_gflags==3.0.6
